remove trainer from megatronbertmodule
